pyproject.toml
[tool.poetry]
name = "nexa-connect-exaa"
version = "0.0.0b1"
description = "Python client for the EXAA Trading API. Day-ahead auction order management, results retrieval, and state polling for Austrian, German, and Dutch power markets."
authors = ["Phase Nexa <hello@phasenexa.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/phasenexa/nexa-connect-exaa"
repository = "https://github.com/phasenexa/nexa-connect-exaa"
keywords = ["energy-trading", "exaa", "day-ahead", "power-market", "electricity-market", "python", "trading-api", "auction", "market-coupling", "european-energy", "nordpool", "epex", "entso-e", "pandas", "rest-api", "pydantic"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [{include = "nexa_connect_exaa", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.0"
httpx = "^0.27"
cryptography = "^42.0"
PyJWT = {version = "^2.8", extras = ["cryptography"]}
pandas = {version = "^2.0", optional = true}

[tool.poetry.extras]
pandas = ["pandas"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
pytest-asyncio = "^0.23"
respx = "^0.21"
ruff = "^0.4"
mypy = "^1.10"
nbmake = "^1.5"
jupyter = "^1.0"
pandas = "^2.0"
pandas-stubs = "^2.0"
cryptography = "^42.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "ANN", "B", "SIM"]
ignore = [
    "ANN101",  # Missing type annotation for `self`
    "ANN102",  # Missing type annotation for `cls`
    "ANN401",  # typing.Any in return types (acceptable for HTTP/JSON response parsing)
]

[tool.mypy]
strict = true
python_version = "3.11"
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=nexa_connect_exaa --cov-report=term-missing --cov-fail-under=80"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/nexa_connect_exaa"]
omit = ["*/pandas_helpers.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
